refactor: split on modules
