+++ english-for-kids/src/helpers/constants.js
@@ -1,0 +1,38 @@
+const classHide = 'hide';
+const startPage = 'Categories';
+
+const EFFECTS = {
+  correct: '/assets/audio/correct.mp3',
+  error: '/assets/audio/error.mp3',
+  failure: '/assets/audio/failure.mp3',
+  success: '/assets/audio/success.mp3',
+};
+
+const MODES = {
+  train: 'train',
+  play: 'play',
+};
+
+const sidenav = document.getElementById('sidenav');
+const logo = document.getElementById('logo');
+const modeBtn = document.getElementById('mode-btn');
+const rating = document.getElementById('rating');
+const pageContainer = document.getElementById('page-container');
+const playBtn = document.getElementById('play-btn');
+const repeatBtn = document.getElementById('repeat-btn');
+const modal = document.getElementById('modal');
+
+export {
+  classHide,
+  startPage,
+  EFFECTS,
+  MODES,
+  sidenav,
+  logo,
+  modeBtn,
+  rating,
+  pageContainer,
+  playBtn,
+  repeatBtn,
+  modal,
+};
